+++ CPP/Tree (Visitable - Composite Component)/IMathematicalOperation.cs
@@ -11,6 +11,6 @@ namespace CPP
     {
         void Evaluate(IVisitor visitor);
 
-        double Data { get; set; }
+        decimal Data { get; set; }
     }
 }
